backend/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.36
asyncpg==0.30.0
greenlet==3.1.1
pydantic==2.11.7
pydantic-settings==2.2.1
python-dotenv==1.0.1
passlib==1.7.4
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
Jinja2==3.1.4
websockets==12.0
email-validator==2.2.0
